requirements-fastapi.txt
# Dumont Cloud - FastAPI Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Requests
requests>=2.28.0
httpx>=0.26.0

# Environment
python-dotenv>=1.0.0

# Previous dependencies (from Flask)
flask>=2.0.0
flask-cors>=4.0.0
